apfs: unify header files
Keeping a header file for each source file doesn't seem to be common
among in-tree filesystems, and I gather that it's disliked by the kernel
developers.  I tried to make it work for apfs, but it doesn't make much
sense because we often require almost every header.

So, unify most of them into a single apfs.h file.  The main exception
is apfs_raw.h: on-disk structures need to remain separate so that they
will never be modified by mistake.  Also ignore unicode.h for now, since
it doesn't quite fit in with the others and might go away when we start
using the official unicode implementation.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>