apfs: improve handling of node subtypes
The subtype for a new node is currently chosen based on the storage
type.  This is of course only possible if we have a single subtype for
each storage type, which will no longer be true once we start working
with physical extents.

To prepare for this, start reading the subtype of nodes into memory on
apfs_read_node(), and give that same subtype to child or sibling nodes
on creation.

The subtype is technically an object property, but only nodes seem to
use it, so place it in the node structure.  Since this is shared by all
nodes in a tree, it may be a good idea to use in-memory tree structures
for this sort of thing in the future.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>