apfs: simplify commit of transaction buffers
Make apfs_transaction_commit() sync all buffers as it iterates through
the list, instead of just marking them as dirty and leaving them for
blkdev_issue_flush().  This change makes it easier to handle objects to
which we still hold references, such as the volume.  It also makes the
commit more reliable: for reasons I ignore, it prevents file buffers
from being lost when we try to implement file writes.

Signed-off-by: Ernesto A. Fernández <ernesto.mnd.fernandez@gmail.com>