apfs: provide basic support for writing to files
Begin work to support file writes.  To this end, add functions to write
file extent records and to deal with the new data stream record type.
A whole new b-tree needs to be used as well: the extent reference tree.
Adjust the node and query code to deal with it.

There is too much I don't understand about the vfs, and writes look
tricky, so try to keep things simple for now.  Don't allow writes
through mmap(), only create extents with a single block (even if this
defeats the whole purpose of extents), and neglect to handle short
writes correctly.  Don't implement any form of truncation yet, and just
let apfs_delete_inode() fail when the inode is not empty.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>