apfs: allow mount of versioned space managers
The versioned space manager is not documented, and I don't know what
it's about; but it's enabled by default in images created by an iphone 6
vm, so I can't ignore it. I don't see any obvious problems so far if I
just allow such images to mount, so do that and hope for the best. Keep
a warning in place to make sure I don't forget to figure this out in the
future.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>