apfs: use the right free queue for ip blocks
My module and userland tools don't use the internal pool for anything,
so when a block is freed it gets put in the main free queue. The problem
is that, if the filesystem image was not created by me, ip blocks may
actually be in use. When those are freed, put them in their proper
queue.

For the record, this is the error message the Apple fsck throws due to
this bug:

  ** Checking the space manager free queue trees.
  error: sfqe_entry : (range 0x1257+0x1, xid 0x3) : Range in free queue tree lies in the internal pool
  error: Spaceman free queue tree of type [1] is invalid
     Space manager free queue trees are invalid.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>