apfs: use internal pool for spaceman structures
The Apple driver has trouble mounting containers that have been touched
by my module, throwing many error messages such as

  spaceman_freed:4137: failed to free internal pool block 10273: 22

Stan figured out that their driver expects cibs, cabs and allocation
bitmaps themselves to all be allocated in a special area called the
internal pool. This makes sense: if these structures are allocated like
any other, you have to take special measures to prevent recursive calls,
such as allocating a new block in the same chunk as its bitmap and cib.
This could become harder as the disk fills and becomes fragmented.

So do as Apple expects and use the internal pool for this stuff. With
this patch my module passes the new checks added to my fsck, but I'm yet
to test if it mounts well in a mac. All problems are not solved, since I
haven't dealt with the ip_bm_free_next field.

Signed-off-by: Ernesto A. Fernández <ernesto@corellium.com>